Move Communicator outside of decision_scene.
